+++ src/main/java/instagram/config/WebSecurityConfig.java
@@ -20,8 +20,6 @@ import javax.sql.DataSource;
 @EnableWebSecurity
 public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
     @Autowired
-    private DataSource dataSource;
-    @Autowired
     private UserService userService;
     @Override
     protected void configure(HttpSecurity http) throws Exception {
